refactor : 이미지 상대경로를 절대경로로 컨버팅 하는 기능을 util/ImagePathConverter로 분리 (#37)
